[COT-12082202] Implement cyber installer update manager, view and logic
_Implement task handle manager for managing singleton task
_Implement new feature: update cyber_installer directly via button click

+++ CyberTool/cyber_installer_background_service/Program.cs
@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics;
+using System.IO.Compression;
 
 namespace cyber_installer_background_service
 {
@@ -6,18 +7,14 @@ namespace cyber_installer_background_service
     {
         private const string CYBER_INSTALLER_REQUESTER_ID = "CyberInstallerWindow{0367E847-B5C3-4CDD-9C34-B78A769AF73C}";
         private const string UPDATE_CYBER_INSTALLER_CMD = "UpdateCyberInstaller";
+        private const string CIBS_FOLDER_ZIP_PATH = "cibs";
 
         static void Main(string[] args)
         {
-            Console.ReadKey();
-            foreach (var arg in args)
-            {
-                Console.WriteLine(arg);
-            }
             var requesterID = "";
             var requesterProcessId = -1;
             var cmdID = "";
-            if (args.Length == 3)
+            if (args.Length >= 3)
             {
                 requesterID = args[0];
                 cmdID = args[1];
@@ -26,23 +23,44 @@ namespace cyber_installer_background_service
                 {
                     case CYBER_INSTALLER_REQUESTER_ID:
                         {
-                            if (cmdID == UPDATE_CYBER_INSTALLER_CMD)
+                            if (cmdID == UPDATE_CYBER_INSTALLER_CMD && args.Length == 5)
                             {
-
-                                try
+                                var zipFilePath = args[3].ToString();
+                                var installFolderLocation = args[4].ToString();
+                                if (File.Exists(zipFilePath) && Directory.Exists(installFolderLocation))
                                 {
-                                    var requesterProcess = Process.GetProcessById(requesterProcessId);
-                                    if (!requesterProcess.HasExited && requesterProcess?.ProcessName == "Cyber Installer")
+                                    try
                                     {
-                                        requesterProcess?.Kill();
-                                        requesterProcess?.WaitForExit();
+                                        var requesterProcess = Process.GetProcessById(requesterProcessId);
+                                        if (!requesterProcess.HasExited && requesterProcess?.ProcessName == "Cyber Installer")
+                                        {
+                                            requesterProcess?.Kill();
+                                            requesterProcess?.WaitForExit();
+                                        }
+
+                                        using (ZipArchive archive = ZipFile.OpenRead(zipFilePath))
+                                        {
+                                            foreach (ZipArchiveEntry entry in archive.Entries)
+                                            {
+                                                if (Path.GetDirectoryName(entry.FullName) != CIBS_FOLDER_ZIP_PATH)
+                                                {
+                                                    entry.ExtractToFile(installFolderLocation
+                                                        + "\\" + Path.GetFileName(entry.FullName)
+                                                        , overwrite: true);
+                                                }
+                                            }
+                                        }
+                                    }
+                                    catch
+                                    {
+
                                     }
                                 }
-                                catch
+                                else
                                 {
-
+                                    Console.WriteLine("Install file not found, please retry!!");
                                 }
-                                
+
                             }
                             break;
                         }
@@ -52,5 +70,7 @@ namespace cyber_installer_background_service
                 Console.ReadKey();
             }
         }
+
+
     }
 }
